+++ CHANGELOG.md
@@ -84,6 +84,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add groups to user type - #5368 by @IKarbowiak
 - Allow specifying users in groups mutations - #5362 by @IKarbowiak
 - Add user can manage field to group type - #5376 by @IKarbowiak
+- Prevent users from modifying a group with permissions they don't hold - #5380 by @IKarbowiak
 
 ## 2.9.0